% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klasifikasi.R
\name{build}
\alias{build}
\title{Build package}
\usage{
build(clients_data, cfg = config())
}
\arguments{
\item{cfg}{list of package options, created by created by [config()]}

\item{client_data}{Nested list with client_id & client_secret key}
}
\value{
List of client_data, client_auth & client_model with publicId key
}
\description{
Retrieve client token & client model data
}
