% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klasifikasi.R
\name{classify}
\alias{classify}
\title{Classyfing into a model}
\usage{
classify(public_id, client_list, query, cfg = config())
}
\arguments{
\item{public_id}{public id of a model, you can get it from your model page in klasifikasi webapp}

\item{client_list}{client list from a model, you can get it from [build()[['publicId']]]}

\item{cfg}{list of package options, created by created by [config()]}
}
\value{
list of tag result
}
\description{
Get classifying result from a model
}
