% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_classifying_result}
\alias{get_classifying_result}
\title{Get classify result}
\usage{
get_classifying_result(client_token, public_id, query, cfg)
}
\arguments{
\item{client_token}{client token, created by [build()[['publicId']]$client_auth$token]}

\item{public_id}{public id model}

\item{query}{query}

\item{cfg}{list of package options, created by [config()]}
}
\description{
Use this function to get classifying result
}
