% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_logs}
\alias{get_logs}
\title{Get model logs}
\usage{
get_logs(client_token, public_id, list_query, cfg)
}
\arguments{
\item{client_token}{client token, created by [build()[['publicId']]$client_auth$token]}

\item{public_id}{public id model}

\item{cfg}{list of package options, created by [config()]}

\item{query_string}{list of query string in url}
}
\description{
Use this function get classifying logs from a model
}
