% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{get_model_data}
\alias{get_model_data}
\title{Get model data}
\usage{
get_model_data(client_token, cfg)
}
\arguments{
\item{client_token}{client token, created by [request_token()$body]}

\item{cfg}{list of package options, created by [config()]}
}
\description{
Use this function to get model data based on client token
}
