% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klasifikasi.R
\name{logs}
\alias{logs}
\title{Logs a model}
\usage{
logs(
  public_id,
  client_list,
  started_at,
  ended_at,
  skip = 0,
  take = 10,
  cfg = config()
)
}
\arguments{
\item{public_id}{public id of a model, you can get it from your model page in klasifikasi webapp}

\item{client_list}{client list from a model, you can get it from [build()[['publicId']]]}

\item{started_at}{Date with 8601 format}

\item{ended_at}{Date with 8601 format}

\item{cfg}{list of package options, created by created by [config()]}
}
\value{
list of model logs
}
\description{
Get logs from a model
}
