% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_token}
\alias{request_token}
\title{Request client token}
\usage{
request_token(client_data, cfg)
}
\arguments{
\item{client_data}{list of clientId & clientSecret}

\item{cfg}{list of package options, created by created by [config()]}
}
\description{
Use this function to request client data
}
